Update README.md
OBJECTIVE: The primary goal of this dashboard is to provide insights into attritions in the company. Insights through this dashboard can help the company to gather all the major KPIs related to employee attrition which can help to have a more employee-centric workplace.

KEY INSIGHTS:

The age group of 26 to 35 experienced the highest attrition rate.

Individuals with a background in Life Sciences and those in Laboratory Technician roles had the highest attrition percentages.

Employees with 1 to 2 years of experience showed the highest attrition rates.

Job satisfaction was found to be directly proportional to attrition.

Employees earning below the average salary of 6.5k were more likely to leave the company.